Control.Monad.Perhaps

Description

Maybe with an undisclosed error

data Perhaps a

This monad occupies the middle ground between Maybe and Either in that you can get out an informative error but aren't able to care about its contents, except via bottoms.

Since bottoms are indistinguishable in pure code, one can view this as morally the same as Maybe, except when things go wrong, you can pass along a complaint, rather than take what you'd get from fromJust.

>>> import Control.Exception
>>> let x = excuse Overflow :: Perhaps ()

Attempting to Show a Perhaps value is hazardous, as it will contain an embedded exception.

>>> x
Can't *** Exception: arithmetic overflow

Recovery is possible as 'Can\'t' isn't strict in its argument.

>>> x <|> Can ()
Can ()

>>> x `seq` ()
()

Constructors

Can a

Can't Void

believe :: Perhaps a -> a

This partial function can be used like fromJust, but throws the user error.

mayhap :: Perhaps a -> Maybe a

Transformer

newtype PerhapsT m a

Constructors

PerhapsT

Fields

runPerhapsT :: m (Perhaps a)

Class

class MonadPlus m => MonadPerhaps m where

Methods

perhaps :: Perhaps a -> m a

This is a monad homomorphism

perhaps :: (m ~ t n, MonadTrans t, MonadPerhaps n) => Perhaps a -> m a

This is a monad homomorphism

excuse :: Exception e => e -> m a

Fail with an exception as an excuse instead of just a string.

Combinators

mapPerhapsT :: (m (Perhaps a) -> n (Perhaps b)) -> PerhapsT m a -> PerhapsT n b

Transform the computation inside a PerhapsT.

runPerhapsT (mapPerhapsT f m) = f (runPerhapsT m)

liftCallCC :: CallCC m (Perhaps a) (Perhaps b) -> CallCC (PerhapsT m) a b

Lift a callCC operation to the new monad.

liftCatch :: Catch e m (Perhaps a) -> Catch e (PerhapsT m) a

Lift a catchE operation to the new monad.

liftListen :: Monad m => Listen w m (Perhaps a) -> Listen w (PerhapsT m) a

Lift a listen operation to the new monad.

liftPass :: Monad m => Pass w m (Perhaps a) -> Pass w (PerhapsT m) a

Lift a pass operation to the new monad.